requirements.txt
fastapi==0.115.5
openai==1.55.0
playwright==1.41.1
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.32.1
